Split repositories into partial files
And adjust some XML comments
