+++ src/AppLibrary/Domain/Repositories/EFRepository/BaseRepository.Exists.cs
@@ -1,0 +1,26 @@
+using GaEpd.AppLibrary.Domain.Entities;
+using Microsoft.EntityFrameworkCore;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq.Expressions;
+
+namespace GaEpd.AppLibrary.Domain.Repositories.EFRepository;
+
+/// <summary>
+/// An implementation of <see cref="IRepository{TEntity,TKey}"/> using Entity Framework.
+/// </summary>
+/// <typeparam name="TEntity">The entity type.</typeparam>
+/// <typeparam name="TKey">The primary key type for the entity.</typeparam>
+/// <typeparam name="TContext">The type of the <see cref="DbContext"/>.</typeparam>
+[SuppressMessage("", "S2436")]
+public abstract partial class BaseRepository<TEntity, TKey, TContext>
+    where TEntity : class, IEntity<TKey>
+    where TKey : IEquatable<TKey>
+    where TContext : DbContext
+{
+    // ExistsAsync
+    public Task<bool> ExistsAsync(TKey id, CancellationToken token = default) =>
+        ExistsAsync(entity => entity.Id.Equals(id), token);
+
+    public async Task<bool> ExistsAsync(Expression<Func<TEntity, bool>> predicate, CancellationToken token = default) =>
+        await NoTrackingSet().AnyAsync(predicate, token).ConfigureAwait(false);
+}
