+++ src/AppLibrary/Domain/Repositories/IReadRepository.Find.cs
@@ -1,0 +1,49 @@
+using System.Linq.Expressions;
+
+namespace GaEpd.AppLibrary.Domain.Repositories;
+
+public partial interface IReadRepository<TEntity, in TKey>
+{
+    /// <summary>
+    /// Returns the <see cref="TEntity"/> matching the given <paramref name="id"/>.
+    /// Returns null if there are no matches.
+    /// </summary>
+    /// <param name="id">The ID of the entity.</param>
+    /// <param name="token"><see cref="T:System.Threading.CancellationToken"/></param>
+    /// <exception cref="InvalidOperationException">Thrown if there are multiple matches.</exception>
+    /// <returns>An entity or null.</returns>
+    Task<TEntity?> FindAsync(TKey id, CancellationToken token = default);
+
+    /// <summary>
+    /// Returns the <see cref="TEntity"/> matching the given <paramref name="id"/>.
+    /// Returns null if there are no matches.
+    /// </summary>
+    /// <param name="id">The ID of the entity.</param>
+    /// <param name="includeProperties">navigation properties to include (when using an Entity Framework repository).</param>
+    /// <param name="token"><see cref="T:System.Threading.CancellationToken"/></param>
+    /// <exception cref="InvalidOperationException">Thrown if there are multiple matches.</exception>
+    /// <returns>An entity or null.</returns>
+    Task<TEntity?> FindAsync(TKey id, string[] includeProperties, CancellationToken token = default);
+
+    /// <summary>
+    /// Returns a single <see cref="TEntity"/> matching the conditions of the <paramref name="predicate"/>.
+    /// Returns null if there are no matches.
+    /// </summary>
+    /// <param name="predicate">The search conditions.</param>
+    /// <param name="token"><see cref="T:System.Threading.CancellationToken"/></param>
+    /// <exception cref="InvalidOperationException">Thrown if there are multiple matches.</exception>
+    /// <returns>An entity or null.</returns>
+    Task<TEntity?> FindAsync(Expression<Func<TEntity, bool>> predicate, CancellationToken token = default);
+
+    /// <summary>
+    /// Returns a single <see cref="TEntity"/> matching the conditions of the <paramref name="predicate"/>.
+    /// Returns null if there are no matches.
+    /// </summary>
+    /// <param name="predicate">The search conditions.</param>
+    /// <param name="includeProperties">Navigation properties to include (when using an Entity Framework repository).</param>
+    /// <param name="token"><see cref="T:System.Threading.CancellationToken"/></param>
+    /// <exception cref="InvalidOperationException">Thrown if there are multiple matches.</exception>
+    /// <returns>An entity or null.</returns>
+    Task<TEntity?> FindAsync(Expression<Func<TEntity, bool>> predicate, string[] includeProperties,
+        CancellationToken token = default);
+}
