+++ src/AppLibrary/Domain/Repositories/LocalRepository/BaseRepository.Exists.cs
@@ -1,0 +1,23 @@
+using GaEpd.AppLibrary.Domain.Entities;
+using System.Linq.Expressions;
+
+namespace GaEpd.AppLibrary.Domain.Repositories.LocalRepository;
+
+/// <summary>
+/// An implementation of <see cref="IRepository{TEntity,TKey}"/> using in-memory data.
+/// </summary>
+/// <typeparam name="TEntity">The entity type.</typeparam>
+/// <typeparam name="TKey">The primary key type for the entity.</typeparam>
+/// <remarks>Navigation properties are already included when using in-memory data structures,
+/// so any `includeProperties` parameters are ignored.</remarks>
+public abstract partial class BaseRepository<TEntity, TKey>
+    where TEntity : IEntity<TKey>
+    where TKey : IEquatable<TKey>
+{
+    // ExistsAsync
+    public Task<bool> ExistsAsync(TKey id, CancellationToken token = default) =>
+        ExistsAsync(entity => entity.Id.Equals(id), token);
+
+    public async Task<bool> ExistsAsync(Expression<Func<TEntity, bool>> predicate, CancellationToken token = default) =>
+        await Task.FromResult(Items.Any(predicate.Compile())).ConfigureAwait(false);
+}
