+++ src/AppLibrary/Domain/Repositories/LocalRepository/BaseRepository.Write.cs
@@ -1,0 +1,32 @@
+using GaEpd.AppLibrary.Domain.Entities;
+
+namespace GaEpd.AppLibrary.Domain.Repositories.LocalRepository;
+
+/// <summary>
+/// An implementation of <see cref="IRepository{TEntity,TKey}"/> using in-memory data.
+/// </summary>
+/// <typeparam name="TEntity">The entity type.</typeparam>
+/// <typeparam name="TKey">The primary key type for the entity.</typeparam>
+/// <remarks>Navigation properties are already included when using in-memory data structures,
+/// so any `includeProperties` parameters are ignored.</remarks>
+public abstract partial class BaseRepository<TEntity, TKey>
+    where TEntity : IEntity<TKey>
+    where TKey : IEquatable<TKey>
+{
+    // IWriteRepository methods
+    public Task InsertAsync(TEntity entity, bool autoSave = true, CancellationToken token = default)
+    {
+        Items.Add(entity);
+        return Task.CompletedTask;
+    }
+
+    public async Task UpdateAsync(TEntity entity, bool autoSave = true, CancellationToken token = default)
+    {
+        var item = await GetAsync(entity.Id, token: token).ConfigureAwait(false);
+        Items.Remove(item);
+        Items.Add(entity);
+    }
+
+    public async Task DeleteAsync(TEntity entity, bool autoSave = true, CancellationToken token = default) =>
+        Items.Remove(await GetAsync(entity.Id, token: token).ConfigureAwait(false));
+}
